moved helper functions to seperate file to clean things up
